+++ src/metascraper.js
@@ -1,5 +1,5 @@
 ﻿/*!
- * Metascraper JavaScript Library v0.6.2
+ * Metascraper JavaScript Library v0.7.0
  * https://metascraper.com/
  *
  * Copyright Metascraper
@@ -13,14 +13,15 @@ if ("undefined" == typeof jQuery) throw new Error("Metascraper requires jQuery")
 
 var meta = {
 
+    //_self: this,
 
     projectName: "Metascraper",
-    version: "0.6.2",
+    version: "0.7.0",
 
     options: {
-        webApiPrefix: "http://localhost/",	    	// fully qualified webapi root
+        webApiPrefix: "http://localhost:54702/api/",    // fully qualified webapi root
         UnauthRedir: "/app/login.htm",                  // redirect on 401 Unauthorized
-        attrName: "data-value",                         // property that is the value (key) of the element, sent on page scrape
+        attrName: "data-value",                         // property that is the value (id/key) of the element, sent on page scrape
         attrMemberName: "data-member",                  // name of the data to use, can be blank for json root 
         attrArrayName: "data-array",                    // array to use to populate multi-elements 
         attrTextName: "data-text",                      // array - display value
@@ -44,23 +45,35 @@ var meta = {
     }),
 
     storeTypesEnum: Object.freeze({
-        OneTimeSingleTab: 0,                            // (DEFAULT) Deletes after first read, does not go across tabs
-        OneTimeAllTab: 1,                               // Deletes after first read, can be read by any tabs
-        SessionSingleTab: 2,                            // Stays for entire session, does not go across tabs
-        //SessionAllTab: 3,
-        ForeverAllTab: 4                                // Stays for entire session, can be read by any tabs
+        ExpiryNoExtend: 1,      // extends the expiry each time it is read, 30 days from set
+        ExpiryExtend: 2,        // keep until expiry is met, extending each time it is read, 30 days since last get
+        OneTimeRead: 3,         // allows it to be read only once
+        Forever: 4,             // DEFAULT - keeps until user deletes it
+        Session: 5              // current session, does not go across tabs
     }),
 
     // enum to determine type of data expected to be in an element
     dataTypesEnum: Object.freeze({
-        date: "date",
-        time: "time",
-        datetime: "datetime",
+        date: "date",                   // 12/25/1990 or 2/15/2019
+        time: "time",                   // 1:15:30 AM
+        datetime: "datetime",           // 12/25/2018, 1:15:30 AM
+        month: "month",                 // January
+        month1: "month1",               // J
+        month3: "month3",               // Jan
+        monthNum: "monthNum",           // 1 index based month, January = 1
+        monthNumPad: "monthNumPad",     // 1 index based month padded, January = 01
+        day: "day",                     // Monday
+        day1: "day1",                   // M
+        day2: "day2",                   // Mo
+        day3: "day3",                   // Mon
+        dayNum: "dayNum",               // 1 index based day, 2 in 1/2/2003
+        dayNumPad: "dayNumPad",         // 1 index based day padded, 02 in 1/2/2003
+        year: "year",                   // 1990 or 2020
+        year2: "year2",                 // 90 or 20 - always 2 digit year, even 100-999 and 10000+
         string: "string",
         number: "number",
         json: "json"
     }),
-
 
     // START - RestApi functions
 
@@ -121,21 +134,6 @@ var meta = {
         this.ajax(o);
     },
 
-    token: function () {
-        var store = window.localStorage;
-        return store.getItem(meta.options.tokenName);
-    },
-
-    tokenize: function (opts) {
-        var token = meta.token();
-        if (!(token))
-            return;
-        var authHeader = { 'Authorization': 'Bearer ' + token };
-        if (!(opts.headers))
-            opts.headers = {};
-        this.extend(opts.headers, authHeader);
-    },
-
     get: function (opts) {
         var d = { method: "GET" };
         this.extend(d, opts);
@@ -173,6 +171,23 @@ var meta = {
         this.put(opts);
     },
 
+
+    // START - security functions
+    
+    token: function () {
+        var store = window.localStorage;
+        return store.getItem(meta.options.tokenName);
+    },
+
+    tokenize: function (opts) {
+        var token = meta.token();
+        if (!(token))
+            return;
+        var authHeader = { 'Authorization': 'Bearer ' + token };
+        if (!(opts.headers))
+            opts.headers = {};
+        this.extend(opts.headers, authHeader);
+    },
 
     login: function (opts) {
         var success = opts.success;
@@ -275,7 +290,15 @@ var meta = {
                 else if (type == meta.dataTypesEnum.number)
                     val = new Number(val);
 
-                else if (type == meta.dataTypesEnum.date || type == meta.dataTypesEnum.time || type == meta.dataTypesEnum.datetime) {
+                else if (type == meta.dataTypesEnum.date || type == meta.dataTypesEnum.time || type == meta.dataTypesEnum.datetime ||
+                        type == meta.dataTypesEnum.month || type == meta.dataTypesEnum.month1 || type == meta.dataTypesEnum.month3 ||
+                        type == meta.dataTypesEnum.monthNum || type == meta.dataTypesEnum.monthNumPad ||
+                        type == meta.dataTypesEnum.day || type == meta.dataTypesEnum.day1 || type == meta.dataTypesEnum.day2 ||
+                        type == meta.dataTypesEnum.day3 || 
+                        type == meta.dataTypesEnum.dayNum || type == meta.dataTypesEnum.dayNumPad || 
+                        type == meta.dataTypesEnum.year || type == meta.dataTypesEnum.year2
+                        )
+                {
                     val = new Date(val);
                     if (type == meta.dataTypesEnum.date)
                         val = val.toLocaleDateString();
@@ -283,6 +306,79 @@ var meta = {
                         val = val.toLocaleTimeString();
                     else if (type == meta.dataTypesEnum.datetime)
                         val = val.toLocaleString();
+                    else if (type == meta.dataTypesEnum.month) 
+                    {
+                        var monthNames = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
+                        val = monthNames[val.getMonth()];
+                    } 
+                    else if (type == meta.dataTypesEnum.month1)
+                    {
+                        var monthNames = ['J', 'F', 'M', 'A', 'M', 'J', 'J', 'A', 'S', 'O', 'N', 'D'];
+                        val = monthNames[val.getMonth()];
+                    } 
+                    else if (type == meta.dataTypesEnum.month3)
+                    {
+                        var monthNames = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+                        val = monthNames[val.getMonth()];
+                    } 
+                    else if (type == meta.dataTypesEnum.monthNum)
+                    {
+                        val = val.getMonth() + 1;
+                    }
+                    else if (type == meta.dataTypesEnum.monthNumPad)
+                    {
+                        var month = val.getMonth() + 1;
+                        val = month < 10 ? '0' + month.toString() : month.toString();
+                    }  
+                    else if (type == meta.dataTypesEnum.day)
+                    {
+                        var dayNames = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
+                        val = dayNames[val.getDay()];
+                    } 
+                    else if (type == meta.dataTypesEnum.day1)
+                    {
+                        var dayNames = ['S', 'M', 'T', 'W', 'T', 'F', 'S'];
+                        val = dayNames[val.getDay()];
+                    } 
+                    else if (type == meta.dataTypesEnum.day2)
+                    {
+                        var dayNames = ['Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa'];
+                        val = dayNames[val.getDay()];
+                    } 
+                    else if (type == meta.dataTypesEnum.day3)
+                    {
+                        var dayNames = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
+                        val = dayNames[val.getDay()];
+                    } 
+                    else if (type == meta.dataTypesEnum.dayNum)
+                    {
+                        val = val.getDate();
+                    }
+                    else if (type == meta.dataTypesEnum.dayNumPad)
+                    {
+                        var day = val.getDate();
+                        val = day < 10 ? '0' + day.toString() : day.toString();
+                    }
+                    else if (type == meta.dataTypesEnum.year)
+                        val = val.getFullYear();
+                    else if (type == meta.dataTypesEnum.year2) {
+                        var year = val.getFullYear();
+                        if (year > -10 && year < 10) {
+                            if (year < 0) {
+                                val = '-0' + (year * -1).toString();
+                            } else {
+                                val = '0' + year.toString();
+                            }
+                        } else if (year < 100 && year > -100) {
+                            val = year;
+                        } else {      // this works for all 3+ digit years 
+                            var yearStr = year.toString();
+                            yearStr = year.substr(yearStr.length - 2, 2);
+                            if (year < 0)
+                                yearStr = '-' + yearStr;
+                            val = parseInt(yearStr);  
+                        } 
+                    }
                 }
 
                 else if (type == meta.dataTypesEnum.json)
@@ -316,17 +412,21 @@ var meta = {
 
     // sets all single value elements (textboxes, checkboxes, etc) to the supplied data values. 
     // does not set multiple value elements, such as table, ul, etc  
-    paint: function (data) {
+    paint: function (data, container) {
+        container = container || document; 
         var attrName = meta.options.attrName;
-        var elems = this.selectAll("[" + attrName + "]");   // scrape
+        var elems = this.selectAll("[" + attrName + "]", container);   // scrape
         for (var i = 0; i < elems.length; i++) {
             // set the field to json dynamically
             var value = null;
             if (elems[i].hasAttribute(meta.options.attrMemberName)) {
                 var member = elems[i].getAttribute(meta.options.attrMemberName)
-                if (member && data[member])
-                    value = data[member][elems[i].getAttribute(attrName)]
-            } 
+                if (member && data[member]) {
+                    value = data[member][elems[i].getAttribute(attrName)];
+				}
+            } else {
+                value = data[elems[i].getAttribute(attrName)];
+			}
             if (!(value == null))
                 meta.setValue(elems[i], value);
         }
@@ -451,9 +551,10 @@ var meta = {
 
     // fills all the multiple value elements (select, table, etc) 
     // does not set any values
-    populate: function (data) {
+    populate: function (data, container) {
+        container = container || document; 
         var attrArrayName = meta.options.attrArrayName;
-        var elems = this.selectAll("[" + attrArrayName + "]");   // scrape
+        var elems = this.selectAll("[" + attrArrayName + "]", container);   // scrape
         for (var i = 0; i < elems.length; i++) {
             // set the field to json dynamically
             var fillData = data[elems[i].getAttribute(attrArrayName)]
@@ -467,10 +568,11 @@ var meta = {
 
     // Full data bind. First populates all multiple item elements such as table or select,
     // then sets values of all elements.
-    bind: function(data) {
-        //meta.template(data);
-        meta.populate(data);
-        meta.paint(data);
+    // container is optional
+    bind: function(data, container) {
+        container = container || document; 
+        meta.populate(data, container);
+        meta.paint(data, container);
     },
 
     // START master page functions
@@ -525,8 +627,81 @@ var meta = {
         this.load(o);
     },
 
-    // START util functions
+    // START store functions
 
+    // type defaults to Forever
+    // expiry defaults to 720 hours (30 days), can be ignored when type is forever
+    // can take objects, such as json
+    store_set: function (name, data, type, expiryhours) {
+        type = type || meta.storeTypesEnum.Forever;
+        expiryhours = expiryhours || 720;
+
+        var dataStr = this.toString(data);
+
+        var now = new Date();
+        var obj = {
+            value: dataStr,
+            type: type,
+            expiry: expiryhours,
+            created: now.getTime(),
+            lastread: now.getTime()
+        }
+        var storage = (type === meta.storeTypesEnum.Session)
+                ? window.sessionStorage
+                : window.localStorage;
+        var objStr = JSON.stringify(obj);
+        storage.setItem(name, objStr);
+    },
+
+    // gets the value if its available, else null
+    store_get: function (name) {
+        // get the value from local storage
+        var storage = window.localStorage;
+        var json = storage.getItem(name);
+        // didnt exist, check session
+        if (!(json)) {
+            storage = window.sessionStorage;
+            json = storage.getItem(name);
+        }
+        // value doesnt exist at all
+        if (!(json)) 
+            return null;
+
+        try {
+            var obj = JSON.parse(json);
+            if (obj.type == meta.storeTypesEnum.Session)
+                return obj.value;
+            else if (obj.type == meta.storeTypesEnum.Forever)
+                return obj.value;
+            else if (obj.type == meta.storeTypesEnum.OneTimeRead) {
+                var value = obj.value;
+                storage.setItem(name, null);
+                return value;
+            } else if (obj.type == meta.storeTypesEnum.ExpiryNoExtend) {
+                var expiry = new Date();
+                expiry.setTime(obj.created);
+                expiry.setHours(expiry.getHours() + obj.expiry);
+                var now = new Date();
+                if (now.getTime() <= expiry.getTime())
+                    return obj.value;
+            } else if (obj.type == meta.storeTypesEnum.ExpiryExtend) {
+                var expiry = new Date();
+                expiry.setTime(obj.lastread);
+                expiry.setHours(expiry.getHours() + obj.expiry);
+                var now = new Date();
+                if (now.getTime() <= expiry.getTime()) {
+                    obj.lastread = now.getTime();
+                    // update the lastread value
+                    storage.setItem(name, JSON.stringify(obj));
+                    return obj.value;
+
+                }
+            }
+        } catch (e) { };
+        return null; 
+    },
+
+    // START util functions
 
     /// filters a json array to where a key matches a value
     /// jsonArray: the json Array you want filtered
@@ -566,23 +741,161 @@ var meta = {
         return '';
     },
 
-
     // returns a unique number, not secure
     unique: function () {
         var d = new Date();
         return d.getTime();
     },
 
-    // START jQuery functions wrappers 
+    // makes a copy of whatever you throw at it
+    // all objects (json, array, html) are deep copies
+    // value copy - null, undefined, number, string, date, 
+    // deep copy - json, array, array of json, html node or element
+    // functions too
+    copy: function(src) {
+        // Handle the 3 simple types or null or undefined
+        if (src === null || typeof src != "object") return src;
 
-    extend: function (target, src) { $.extend(target, src); },
+        // handle html element
+        if (src instanceof Node) {
+            if (src.outerHTML) {    // handle elements                
+                return this.parseHTML(src.outerHTML)
+            } else {                // everything else
+                return src.cloneNode(true);
+            }
+        }
+
+        if (src instanceof NodeList) {
+            var docfrag = document.createDocumentFragment();
+            var len = src.childNodes ? src.childNodes.length : src.length;
+            var orig;
+            var copy;
+            for (var i = 0; i < len; i++) {
+                orig = null;
+                if (src.childNodes && src.childNodes.length && src.childNodes[i])
+                    orig = src.childNodes[i];
+                else if (src.length && src[i])
+                    orig = src[i];
+                if (!(orig)) continue;
+                var copy = this.copy(orig);
+                docfrag.appendChild(copy);
+            }
+            return docfrag;
+        }
+
+        // handle array and json objects
+        if (src.constructor === Array || src.constructor === Object) {
+            try {
+                return JSON.parse(JSON.stringify(src)); 
+            } catch (jsonErr) {
+                console.log('Error: meta.copy: unable to parse source as json')
+            }
+        } 
+
+        // handle date
+        if (src.constructor === Date) {
+            target = new Date();
+            target.setTime(src.getTime());
+            return target;
+        }
+
+        // unknown - fallback
+        return new src.constructor(src);
+    },
+
+    toString: function(src) {
+        if (src === null || typeof src != "object") 
+            return src;
+
+        if (src.constructor === Array || src.constructor === Object) {
+            try {
+                return JSON.stringify(src); 
+            } catch (jsonErr) {
+                console.log('Error: meta.toString: unable to parse source as json')
+            }
+        }
+
+        // handle html element
+        if (src instanceof Node) {
+            if (src.outerHTML)     // handle elements                
+                return src.outerHTML;
+        }
+
+        if (src instanceof NodeList) {
+            var html = '';
+            if (docfrag && docfrag.childNodes && docfrag.childNodes.length && docfrag.childNodes.length > 0) {
+                for (var i = 0; i < docfrag.childNodes.length; i++) {
+                    if (!(docfrag.childNodes[i])) continue;
+                    if (docfrag.childNodes[i].outerHTML) {
+                        html += docfrag.childNodes[i].outerHTML;
+                    } else if (docfrag.childNodes[i].textContent) {
+                        html += docfrag.childNodes[i].textContent;
+                    }
+                }
+            }
+            return html;
+        }
+
+        if (src.constructor === Date) {
+            return src.toString();
+        }
+
+        console.log('Error: meta.toString: unable to parse source');
+
+    },
+
+    // can check numbers, semvers, and most any numeric values with unlimited separators
+    // ex 2 > 1, 2.3.5 > 2.3.4, 2.0 > 1.9.9,  2.3.4.5.6 > 2.3.4.5, 3 > 2.3.5.4.7.8.9, etc
+    versionCheck: function(semVer, requiredSemVer) {
+        if (semVer === requiredSemVer)  // they should be strings
+            return true;
+        
+        semVerArr = semVer.toString().split('.');
+        requiredSemVerArr = requiredSemVer.toString().split('.');
+       // for (var i = 0; i < semVerArr.length; i++) {
+       //     if (requiredSemVerArr.length == i)
+        for (var i = 0; i < requiredSemVerArr.length; i++) {
+            if (semVerArr.length == i)
+                  return false;
+            if (parseInt(semVerArr[i]) > parseInt(requiredSemVerArr[i]))
+                return true; 
+            if (parseInt(semVerArr[i]) < parseInt(requiredSemVerArr[i]))
+                return false; 
+        }
+        return true;
+    },
+
+    // START jQuery light 
+    
+    // deep extend of src into target
+    // need to check values of null, undefined, dates
+    // html nodes and functions are by reference, not by value, so changing the value of one affects the other
+    extend: function (target, src) { 
+        for (var key in src) {
+            var srcVal = src[key];
+            if (srcVal === null || typeof srcVal != 'object' || srcVal instanceof Node) {
+                //console.log('key:' + key);
+                target[key] = srcVal;
+            }
+            else if (src.constructor === Date) {
+                //console.log('date:' + key);
+                var d = new Date();
+                d.setTime(srcVal.getTime());
+                target[key] = d;
+            } else {
+                //console.log('extending:' + key);
+                // create a default value
+                target[key] = Array.isArray(src) ? [] : {};     
+                target[key] = this.extend(target[key], srcVal);
+            }
+        }
+        return target;
+    },
 
     ajax: function (opts) { 
         $.ajax(opts); 
     },
-
-    // START jQuery light
-
+    
     ready: function (f) {
         var state = document.readyState;
         if (state === 'interactive' || state === 'complete')
@@ -590,14 +903,71 @@ var meta = {
         else
             setTimeout(function () { meta.ready(f) }, 100);
     },
-
-    select: function (id) {
+    
+    select: function (id) { 
         return document.getElementById(id); 
-    },  
-            
+    },        
+          
     selectAll: function (filter, container) {
-        if (!container) container = document; 
-        return document.querySelectorAll(filter); 
-    }, 
+        container = container || document; 
+        return container.querySelectorAll(filter); 
+    },
+    
+    parseHTML : (function() {
+        var rxhtmlTag = /<(?!area|br|col|embed|hr|img|input|link|meta|param)(([\w:]+)[^>]*)\/>/gi,
+            rtagName = /<([\w:]+)/,
+            rhtml = /<|&#?\w+;/,
+            wrapMap = {
+                // Support: IE9
+                option: [1, "<select multiple='multiple'>", "</select>"],
+
+                thead: [1, "<table>", "</table>"],
+                col: [2, "<table><colgroup>", "</colgroup></table>"],
+                tr: [2, "<table><tbody>", "</tbody></table>"],
+                td: [3, "<table><tbody><tr>", "</tr></tbody></table>"],
+
+                _default: [0, "", ""]
+            };
+        
+        //
+        // @param {String} html A string containing html
+        // @param {Document} context
+        // 
+        return function parseHTML(html, context) {
+            if (!(html)) return null;
+            context = context || document;
+
+            var tmp, tag, wrap, j,
+                fragment = context.createDocumentFragment();
+
+            if (!rhtml.test(html)) {
+                fragment.appendChild(context.createTextNode(html));
+
+                // Convert html into DOM nodes
+            } else {
+                tmp = fragment.appendChild(context.createElement("div"));
+
+                // Deserialize a standard representation
+                tag = (rtagName.exec(html) || ["", ""])[1].toLowerCase();
+                wrap = wrapMap[tag] || wrapMap._default;
+                tmp.innerHTML = wrap[1] + html.replace(rxhtmlTag, "<$1></$2>") + wrap[2];
+
+                // Descend through wrappers to the right content
+                j = wrap[0];
+                while (j--) {
+                    tmp = tmp.lastChild;
+                }
+
+                // Remove wrappers and append created nodes to fragment
+                fragment.removeChild(fragment.firstChild);
+                while (tmp.firstChild) {
+                    fragment.appendChild(tmp.firstChild);
+                }
+            }
+
+            return fragment;
+        };
+    }()),
+
 
 };